+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - 
 ### Fixed:
 - Added meadow compat
+- Added beachparty compat
 ### Changed:
 -